+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.0.2
+- Error handling if object not available
+- Fixed button tinting on selection
+- Using numToString when saving hotbar
+
 # 1.0.1
 - Added custom object support